requirements.txt
phidata
streamlit
python-dotenv
packaging
duckduckgo-search
fastapi
uvicorn
groq
wikipedia
googlesearch-python
pycountry
